client-requirements.txt
requests
flask
flask-socketio
websocket-client